stream-service/code/requirements.txt
flask
tornado
python-dotenv
mysql-connector-python